+++ src/tutorials/hole_punching.rs
@@ -67,28 +67,53 @@
 //! ```
 //!
 //! Now let's make sure that the server is public, in other words let's make sure one can reach it
-//! through the Internet. From the dialing client:
+//! through the Internet. First, either manually replace `$RELAY_SERVER_IP` in the following
+//! commands or `export RELAY_SERVER_IP=ipaddr` with the appropriate relay server `ipaddr` in
+//! the dailing client and listening client.
+//!
+//! Now, from the dialing client:
 //!
 //! 1. Test that you can connect on Layer 3 (IP).
 //!
-//!     ``` bash
-//!     ping $RELAY_SERVER_IP
-//!     ```
+//!    ``` bash
+//!    ping $RELAY_SERVER_IP
+//!    ```
 //!
 //! 2. Test that you can connect on Layer 4 (TCP).
 //!
-//!     ``` bash
-//!     telnet $RELAY_SERVER_IP 4001
-//!     ```
+//!    ``` bash
+//!    telnet $RELAY_SERVER_IP 4001
+//!    ```
 //!
 //! 3. Test that you can connect via libp2p using [`libp2p-lookup`](https://github.com/mxinden/libp2p-lookup).
 //!
-//!     ``` bash
-//!     ## For IPv4
-//!     libp2p-lookup direct --address /ip4/$RELAY_SERVER_IP
-//!     ## For IPv6
-//!     libp2p-lookup direct --address /ip6/$RELAY_SERVER_IP
-//!     ```
+//!    ``` bash
+//!    ## For IPv4
+//!    libp2p-lookup direct --address /ip4/$RELAY_SERVER_IP/tcp/4001
+//!    ## For IPv6
+//!    libp2p-lookup direct --address /ip6/$RELAY_SERVER_IP/tcp/4001
+//!    ```
+//!
+//! The libp2p-lookup output should look something like:
+//!
+//!    ``` bash
+//!    $ libp2p-lookup direct --address /ip4/111.11.111.111/tcp/4001
+//!    Lookup for peer with id PeerId("12D3KooWDpJ7As7BWAwRMfu1VU2WCqNjvq387JEYKDBj4kx6nXTN") succeeded.
+//!
+//!    Protocol version: "/TODO/0.0.1"
+//!    Agent version: "rust-libp2p/0.36.0"
+//!    Observed address: "/ip4/22.222.222.222/tcp/39212"
+//!    Listen addresses:
+//!            - "/ip4/127.0.0.1/tcp/4001"
+//!            - "/ip4/111.11.111.111/tcp/4001"
+//!            - "/ip4/10.48.0.5/tcp/4001"
+//!            - "/ip4/10.124.0.2/tcp/4001"
+//!    Protocols:
+//!            - "/libp2p/circuit/relay/0.2.0/hop"
+//!            - "/ipfs/ping/1.0.0"
+//!            - "/ipfs/id/1.0.0"
+//!            - "/ipfs/id/push/1.0.0"
+//!    ```
 //!
 //! ## Setting up the listening client
 //!
@@ -107,6 +132,14 @@
 //!
 //! ``` bash
 //! RUST_LOG=info ./client --secret-key-seed 1 --mode listen --relay-address /ip4/$RELAY_SERVER_IP/tcp/4001/p2p/12D3KooWDpJ7As7BWAwRMfu1VU2WCqNjvq387JEYKDBj4kx6nXTN
+//!
+//! [2022-05-11T10:38:52Z INFO  client] Local peer id: PeerId("XXX")
+//! [2022-05-11T10:38:52Z INFO  client] Listening on "/ip4/127.0.0.1/tcp/44703"
+//! [2022-05-11T10:38:52Z INFO  client] Listening on "/ip4/XXX/tcp/44703"
+//! [2022-05-11T10:38:54Z INFO  client] Relay told us our public address: "/ip4/XXX/tcp/53160"
+//! [2022-05-11T10:38:54Z INFO  client] Told relay its public address.
+//! [2022-05-11T10:38:54Z INFO  client] Relay accepted our reservation request.
+//! [2022-05-11T10:38:54Z INFO  client] Listening on "/ip4/$RELAY_SERVER_IP/tcp/4001/p2p/12D3KooWDpJ7As7BWAwRMfu1VU2WCqNjvq387JEYKDBj4kx6nXTN/p2p-circuit/p2p/XXX"
 //! ```
 //!
 //! Now let's make sure that the listening client is not public, in other words let's make sure one
@@ -114,7 +147,7 @@
 //! connect on Layer 4 (TCP):
 //!
 //! ``` bash
-//! telnet $RELAY_SERVER_IP 4001
+//! telnet $LISTENING_CLIENT_IP_OBSERVED_BY_RELAY 53160
 //! ```
 //!
 //! ## Connecting to the listening client from the dialing client