+++ swarm/src/connection/pool.rs
@@ -87,6 +87,11 @@ where
     /// The configured override for substream protocol upgrades, if any.
     substream_upgrade_protocol_override: Option<libp2p_core::upgrade::Version>,
 
+    /// The maximum number of inbound streams concurrently negotiating on a connection.
+    ///
+    /// See [`super::handler_wrapper::HandlerWrapper::max_negotiating_inbound_streams`].
+    max_negotiating_inbound_streams: usize,
+
     /// The executor to use for running the background tasks. If `None`,
     /// the tasks are kept in `local_spawns` instead and polled on the
     /// current thread when the [`Pool`] is polled for new events.
@@ -155,13 +160,16 @@ impl<THandler: IntoConnectionHandler, TTrans: Transport> fmt::Debug for Pool<THa
 }
 
 /// Event that can happen on the `Pool`.
-pub enum PoolEvent<'a, THandler: IntoConnectionHandler, TTrans>
+#[derive(Debug)]
+pub enum PoolEvent<THandler: IntoConnectionHandler, TTrans>
 where
     TTrans: Transport,
 {
     /// A new connection has been established.
     ConnectionEstablished {
-        connection: EstablishedConnection<'a, <THandler::Handler as ConnectionHandler>::InEvent>,
+        id: ConnectionId,
+        peer_id: PeerId,
+        endpoint: ConnectedPoint,
         /// List of other connections to the same peer.
         ///
         /// Note: Does not include the connection reported through this event.
@@ -190,8 +198,6 @@ where
         /// The error that occurred, if any. If `None`, the connection
         /// was closed by the local peer.
         error: Option<ConnectionError<<THandler::Handler as ConnectionHandler>::Error>>,
-        /// A reference to the pool that used to manage the connection.
-        pool: &'a mut Pool<THandler, TTrans>,
         /// The remaining established connections to the same peer.
         remaining_established_connection_ids: Vec<ConnectionId>,
         handler: THandler::Handler,
@@ -225,88 +231,21 @@ where
 
     /// A node has produced an event.
     ConnectionEvent {
-        /// The connection that has generated the event.
-        connection: EstablishedConnection<'a, THandlerInEvent<THandler>>,
+        id: ConnectionId,
+        peer_id: PeerId,
         /// The produced event.
         event: THandlerOutEvent<THandler>,
     },
 
     /// The connection to a node has changed its address.
     AddressChange {
-        /// The connection that has changed address.
-        connection: EstablishedConnection<'a, THandlerInEvent<THandler>>,
+        id: ConnectionId,
+        peer_id: PeerId,
         /// The new endpoint.
         new_endpoint: ConnectedPoint,
         /// The old endpoint.
         old_endpoint: ConnectedPoint,
     },
-}
-
-impl<'a, THandler: IntoConnectionHandler, TTrans> fmt::Debug for PoolEvent<'a, THandler, TTrans>
-where
-    TTrans: Transport,
-    TTrans::Error: fmt::Debug,
-{
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
-        match self {
-            PoolEvent::ConnectionEstablished {
-                connection,
-                concurrent_dial_errors,
-                ..
-            } => f
-                .debug_tuple("PoolEvent::ConnectionEstablished")
-                .field(connection)
-                .field(concurrent_dial_errors)
-                .finish(),
-            PoolEvent::ConnectionClosed {
-                id,
-                connected,
-                error,
-                ..
-            } => f
-                .debug_struct("PoolEvent::ConnectionClosed")
-                .field("id", id)
-                .field("connected", connected)
-                .field("error", error)
-                .finish(),
-            PoolEvent::PendingOutboundConnectionError {
-                id, error, peer, ..
-            } => f
-                .debug_struct("PoolEvent::PendingOutboundConnectionError")
-                .field("id", id)
-                .field("error", error)
-                .field("peer", peer)
-                .finish(),
-            PoolEvent::PendingInboundConnectionError {
-                id,
-                error,
-                send_back_addr,
-                local_addr,
-                ..
-            } => f
-                .debug_struct("PoolEvent::PendingInboundConnectionError")
-                .field("id", id)
-                .field("error", error)
-                .field("send_back_addr", send_back_addr)
-                .field("local_addr", local_addr)
-                .finish(),
-            PoolEvent::ConnectionEvent { connection, event } => f
-                .debug_struct("PoolEvent::ConnectionEvent")
-                .field("peer", &connection.peer_id())
-                .field("event", event)
-                .finish(),
-            PoolEvent::AddressChange {
-                connection,
-                new_endpoint,
-                old_endpoint,
-            } => f
-                .debug_struct("PoolEvent::AddressChange")
-                .field("peer", &connection.peer_id())
-                .field("new_endpoint", new_endpoint)
-                .field("old_endpoint", old_endpoint)
-                .finish(),
-        }
-    }
 }
 
 impl<THandler, TTrans> Pool<THandler, TTrans>
@@ -329,6 +268,7 @@ where
             task_command_buffer_size: config.task_command_buffer_size,
             dial_concurrency_factor: config.dial_concurrency_factor,
             substream_upgrade_protocol_override: config.substream_upgrade_protocol_override,
+            max_negotiating_inbound_streams: config.max_negotiating_inbound_streams,
             executor: config.executor,
             local_spawns: FuturesUnordered::new(),
             pending_connection_events_tx,
@@ -477,15 +417,25 @@ where
     /// has been reached.
     pub fn add_outgoing(
         &mut self,
-        transport: TTrans,
-        addresses: impl Iterator<Item = Multiaddr> + Send + 'static,
+        dials: Vec<
+            BoxFuture<
+                'static,
+                (
+                    Multiaddr,
+                    Result<
+                        <TTrans as Transport>::Output,
+                        TransportError<<TTrans as Transport>::Error>,
+                    >,
+                ),
+            >,
+        >,
         peer: Option<PeerId>,
         handler: THandler,
         role_override: Endpoint,
         dial_concurrency_factor_override: Option<NonZeroU8>,
     ) -> Result<ConnectionId, (ConnectionLimit, THandler)>
     where
-        TTrans: Clone + Send,
+        TTrans: Send,
         TTrans::Dial: Send + 'static,
     {
         if let Err(limit) = self.counters.check_max_pending_outgoing() {
@@ -493,11 +443,8 @@ where
         };
 
         let dial = ConcurrentDial::new(
-            transport,
-            peer,
-            addresses,
+            dials,
             dial_concurrency_factor_override.unwrap_or(self.dial_concurrency_factor),
-            role_override,
         );
 
         let connection_id = self.next_connection_id();
@@ -543,7 +490,7 @@ where
     where
         TFut: Future<Output = Result<TTrans::Output, TTrans::Error>> + Send + 'static,
     {
-        let endpoint = info.to_connected_point();
+        let endpoint = info.create_connected_point();
 
         if let Err(limit) = self.counters.check_max_pending_incoming() {
             return Err((limit, handler));
@@ -575,11 +522,9 @@ where
         );
         Ok(connection_id)
     }
+
     /// Polls the connection pool for events.
-    ///
-    /// > **Note**: We use a regular `poll` method instead of implementing `Stream`,
-    /// > because we want the `Pool` to stay borrowed if necessary.
-    pub fn poll<'a>(&'a mut self, cx: &mut Context<'_>) -> Poll<PoolEvent<'a, THandler, TTrans>>
+    pub fn poll(&mut self, cx: &mut Context<'_>) -> Poll<PoolEvent<THandler, TTrans>>
     where
         TTrans: Transport<Output = (PeerId, StreamMuxerBox)>,
         THandler: IntoConnectionHandler + 'static,
@@ -595,16 +540,7 @@ where
             Poll::Ready(None) => unreachable!("Pool holds both sender and receiver."),
 
             Poll::Ready(Some(task::EstablishedConnectionEvent::Notify { id, peer_id, event })) => {
-                let entry = self
-                    .established
-                    .get_mut(&peer_id)
-                    .expect("Receive `Notify` event for established peer.")
-                    .entry(id)
-                    .expect_occupied("Receive `Notify` event from established connection");
-                return Poll::Ready(PoolEvent::ConnectionEvent {
-                    connection: EstablishedConnection { entry },
-                    event,
-                });
+                return Poll::Ready(PoolEvent::ConnectionEvent { peer_id, id, event });
             }
             Poll::Ready(Some(task::EstablishedConnectionEvent::AddressChange {
                 id,
@@ -622,16 +558,12 @@ where
                 let old_endpoint =
                     std::mem::replace(&mut connection.endpoint, new_endpoint.clone());
 
-                match self.get(id) {
-                    Some(PoolConnection::Established(connection)) => {
-                        return Poll::Ready(PoolEvent::AddressChange {
-                            connection,
-                            new_endpoint,
-                            old_endpoint,
-                        })
-                    }
-                    _ => unreachable!("since `entry` is an `EstablishedEntry`."),
-                }
+                return Poll::Ready(PoolEvent::AddressChange {
+                    peer_id,
+                    id,
+                    new_endpoint,
+                    old_endpoint,
+                });
             }
             Poll::Ready(Some(task::EstablishedConnectionEvent::Closed {
                 id,
@@ -656,7 +588,6 @@ where
                     connected: Connected { endpoint, peer_id },
                     error,
                     remaining_established_connection_ids,
-                    pool: self,
                     handler,
                 });
             }
@@ -761,7 +692,7 @@ where
                     if let Err(error) = error {
                         self.spawn(
                             poll_fn(move |cx| {
-                                if let Err(e) = ready!(muxer.close(cx)) {
+                                if let Err(e) = ready!(muxer.poll_close(cx)) {
                                     log::debug!(
                                         "Failed to close connection {:?} to peer {}: {:?}",
                                         id,
@@ -819,6 +750,7 @@ where
                         muxer,
                         handler.into_handler(&obtained_peer_id, &endpoint),
                         self.substream_upgrade_protocol_override,
+                        self.max_negotiating_inbound_streams,
                     );
                     self.spawn(
                         task::new_for_established_connection(
@@ -834,7 +766,9 @@ where
                     match self.get(id) {
                         Some(PoolConnection::Established(connection)) => {
                             return Poll::Ready(PoolEvent::ConnectionEstablished {
-                                connection,
+                                peer_id: connection.peer_id(),
+                                endpoint: connection.endpoint().clone(),
+                                id: connection.id(),
                                 other_established_connection_ids,
                                 concurrent_dial_errors,
                             })
@@ -1226,6 +1160,11 @@ pub struct PoolConfig {
 
     /// The configured override for substream protocol upgrades, if any.
     substream_upgrade_protocol_override: Option<libp2p_core::upgrade::Version>,
+
+    /// The maximum number of inbound streams concurrently negotiating on a connection.
+    ///
+    /// See [super::handler_wrapper::HandlerWrapper::max_negotiating_inbound_streams].
+    max_negotiating_inbound_streams: usize,
 }
 
 impl Default for PoolConfig {
@@ -1237,6 +1176,7 @@ impl Default for PoolConfig {
             // By default, addresses of a single connection attempt are dialed in sequence.
             dial_concurrency_factor: NonZeroU8::new(1).expect("1 > 0"),
             substream_upgrade_protocol_override: None,
+            max_negotiating_inbound_streams: 128,
         }
     }
 }
@@ -1293,6 +1233,14 @@ impl PoolConfig {
         v: libp2p_core::upgrade::Version,
     ) -> Self {
         self.substream_upgrade_protocol_override = Some(v);
+        self
+    }
+
+    /// The maximum number of inbound streams concurrently negotiating on a connection.
+    ///
+    /// See [`super::handler_wrapper::HandlerWrapper::max_negotiating_inbound_streams`].
+    pub fn with_max_negotiating_inbound_streams(mut self, v: usize) -> Self {
+        self.max_negotiating_inbound_streams = v;
         self
     }
 }