+++ core/CHANGELOG.md
@@ -1,3 +1,22 @@
+# 0.33.1 - unreleased
+
+- Introduce `StreamMuxerEvent::map_inbound_stream`. See [PR 2691].
+
+[PR 2691]: https://github.com/libp2p/rust-libp2p/pull/2691
+
+# 0.33.0
+
+- Have methods on `Transport` take `&mut self` instead of `self`. See [PR 2529].
+- Remove `StreamMuxer::flush_all`. See [PR 2669].
+- Rename `StreamMuxer::close` to `StreamMuxer::poll_close`. See [PR 2666].
+- Remove deprecated function `StreamMuxer::is_remote_acknowledged`. See [PR 2665].
+
+[PR 2529]: https://github.com/libp2p/rust-libp2p/pull/2529
+[PR 2666]: https://github.com/libp2p/rust-libp2p/pull/2666
+[PR 2665]: https://github.com/libp2p/rust-libp2p/pull/2665
+[PR 2669]: https://github.com/libp2p/rust-libp2p/pull/2669
+
+
 # 0.32.1
 
 - Add `PeerId::try_from_multiaddr` to extract a `PeerId` from a `Multiaddr` that ends in `/p2p/<peer-id>`.